requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyMySQL==0.10.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
virtualenv==20.0.33
Werkzeug==1.0.1
